fix(oracle): widen qtg_applications int->int64 end-to-end (bd 9fi)
`mod->qtg_applications` is the post-parallel oracle-count AGGREGATION slot:
Model.solve() writes `mod.qtg_applications = max(per-worker oracle_count)`
once, single-threaded, after the joblib fan-out (Model.pyx:826), then the
`oracle_calls` property / OptimizeResult read it. The per-worker live counter
moved to a race-free `size_t ctx->oracle_count` (bd lif/8an.1.4); the C hot
loop no longer touches this field (SearchLib.c references it only in comments).

As an `int` (int32) the slot truncates in C and raises OverflowError at the
Cython assignment when a raw-API caller sets a huge `mod->M` and the aggregate
exceeds 2^31-1. Defense-in-depth complement to bd 8an.1.17's cap-j clamp
(which keeps normal-use counts <= T(n) <= INT32_MAX).

Widen to int64_t in model.h AND its Model.pxd mirror in the SAME commit (a
half-applied widening mis-narrows at the Cython boundary). model.h now
`#include <stdint.h>` directly (was transitive via constraint.h) so an include
reorder can't break the declaration. Rebuild regenerates the Cython boundary to
__Pyx_PyInt_As_int64_t (write) / __Pyx_PyInt_From_int64_t (reads). The per-worker
counter, the 2j+1 charge, signs, and the phase machine are all unchanged.

New tests/test_model.c::test_model_qtg_applications_int64 pins the storage width
(sizeof==8) and a >INT32_MAX round-trip — RED on the old `int` field, GREEN on
int64_t.

Vetted by the core-change gate (verdict: sound — faithfulness/threading/sign/
phase/baseline all OK, is_redesign=false). Clean rebuild; C suite 64/64,
Python 671 passed/4 skipped, stress 4 passed; determinism (§8) intact.

Note: a struct-width change shifts later field offsets, so it needs a clean
rebuild (rm -rf build/) — setuptools build_clib doesn't track header deps.

+++ cbqs/Model.pxd
@@ -24,7 +24,7 @@ cdef extern from "src/model.h":
 		int reset_delta;
 		int max_delta;
 		int solver
-		int qtg_applications
+		int64_t qtg_applications   # bd 9fi: int64 to mirror model.h (must match exactly)
 		int max_worse_acceptances;
 		int stopping_condition;
 		int distance;
